Add onSubmit functionality, call api
